Moved output to tests folder, implemented more operations.
The application will now dump the function in a readable format with a signature.

Implemented the following operations:
- Call
- AsObject
- ReturnWithValue

Changes done to op codes:
Any opcode that now uses decodeOperation check if the operation failed and increment to the next op (fail-safe).

Updated README to use generic folder rather then the debug folder.